pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tasknotes-mcp"
version = "0.1.0"
description = "MCP server bridging Claude with the Obsidian TaskNotes HTTP API"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",         # Anthropic MCP Python SDK
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",        # httpx mocking for tests
]

[project.scripts]
tasknotes-mcp = "tasknotes_mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
